Basic permissions (admin/read)

+++ src/routes/applications/[id]/delete.json.ts
@@ -1,9 +1,11 @@
-import { selectTeam } from '$lib/common';
+import { getUserDetails } from '$lib/common';
 import * as db from '$lib/database';
 import type { RequestHandler } from '@sveltejs/kit';
 
 export const del: RequestHandler<Locals, FormData> = async (request) => {
-    const teamId = selectTeam(request)
+	const { teamId, status, body } = await getUserDetails(request);
+	if (status === 401) return { status, body }
+    
     const { id } = request.params
     try {
         await db.prisma.application.deleteMany({ where: { id, teams: { some: { id: teamId } } } })
